requirements.txt
anyio==3.6.1
click==8.1.3
colorama==0.4.5
ecdsa==0.18.0
fastapi==0.79.0
h11==0.13.0
idna==3.3
pyasn1==0.4.8
pydantic==1.9.2
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.9
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
typing_extensions==4.3.0
uvicorn==0.18.2
